requirements.txt
fastapi[all]
python-multipart
python-jose
passlib
boto3
mangum